feat: cart and caffein cache added

+++ active-life-canada/src/main/java/io/reactivestax/active_life_canada/domain/FamilyGroup.java
@@ -19,7 +19,7 @@ public class FamilyGroup {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long familyGroupId;
     private String familyPin;
-    private Integer credits = 0;
+    private Double credits = 0.0;
     private String status = "inActive";
     private String groupOwner;
     private Integer failedLoginAttempts;
